Implemented the tagging function
Implemented the tagging on both strings and token lists.

I also used AI (Calude Sonnet 4.5) this time. We are running out of time and this should have been done long ago, so I needed to speed up the process a bit.

However I checked, tested and manually debugged the code, so there should be no surprises.