Evaluation loop working
Implemented the evaluation loop written by @naeembashir63  on his branch inside the class.

Given that we are in a bit of a hurry, I didn't fully read it or correct it, I am just assuming that it works, because the output looks reasonable. If you know that something in there still needs work, please do tell me!